dkp: git lfs support added, updated changed_files logic for central pipeline (#11)
* dkp: lfs integration, changed_files logic update

* Update initiate-push-to-source.sh

* Update README.md

+++ knowledge-portal/central/initiate-push-to-source.sh
@@ -6,15 +6,15 @@ set -e
 git config --global user.email "PlatformTeam@detecttechnologies.com"
 git config --global user.name "Detect Gitlab Bot"
 
-# changed files can be obtained by git diff between latest commit and unstaged area
-git reset --soft HEAD~1
-changed_files=$(git diff --name-only HEAD)
+# changed files are all files since pipeline successfully ran last time on central
+GITLAB_API="https://gitlab.com/api/v4"
+PIPELINES=$(curl --header "PRIVATE-TOKEN: $API_TOKEN" "$GITLAB_API/projects/$CI_PROJECT_ID/pipelines?scope=finished&status=success" | jq '.[0].sha')
+LAST_PIPELINE_COMMIT=$(echo $PIPELINES | tr -d '"')
+changed_files=$(git diff --name-only $LAST_PIPELINE_COMMIT $CI_COMMIT_SHA)
 
 if [[ $changed_files ]]
 then
     echo "Files changed in knowledge portal: ${changed_files}"
-    # Reverse the git reset and move files to staged as we will need latest files 
-    git add -A && git commit -m "docs:reversing soft reset as now we have the filenames"
 
     # Run python script to get output in format "source_repo_name:filepath_central_gitrepo-->filepath_source_gitrepo"
     output=$(python3 -c "$(curl -fsSL https://github.com/detecttechnologies/Gitlab-CI-CD-Templates/raw/main/knowledge-portal/central/central-pipeline-checks.py)" $changed_files)
@@ -70,8 +70,14 @@ then
         echo "Pushing to ${source_repo} and Creating a draft MR"
         source_repo_path=${source_repo//'_'/'/'}
         cd /root/source/${source_repo}
-        git add -A && git commit -m "doc update from knowledge portal"
-        git push -f -o merge_request.create -o merge_request.draft -o merge_request.title="doc update from knowledge portal" -o merge_request.target=main https://oauth2:${BOT_ACCESS_TOKEN}@gitlab.com/DetectTechnologies/${source_repo_path}.git knowledge-portal_changes
+        git add -A 
+        # Check if there are changes to commit
+        if [ -n "$(git status --porcelain)" ]; then
+            git commit -m "doc update from knowledge portal"
+            git push -f -o merge_request.create -o merge_request.draft -o merge_request.title="doc update from knowledge portal" -o merge_request.target=main https://oauth2:${BOT_ACCESS_TOKEN}@gitlab.com/DetectTechnologies/${source_repo_path}.git knowledge-portal_changes
+        else
+            echo "No changes to commit"
+        fi
         cd -
     done
 fi
